code refractor
