+++ app/components/public/side-menu.js
@@ -1,32 +1,85 @@
 import classic from 'ember-classic-decorator';
-import { computed } from '@ember/object';
+import { action, computed } from '@ember/object';
 import Component from '@ember/component';
+import moment from 'moment';
+import { tagName } from '@ember-decorators/component';
+import { SPEAKERS_FILTER } from 'open-event-frontend/routes/public/speakers';
+import { tracked } from '@glimmer/tracking';
 
 @classic
+@tagName('')
 export default class SideMenu extends Component {
 
-  activeSection = null;
+  @tracked
+  showSpeakers = false;
 
-  activeMenuSection = this.activeSection;
+  @tracked
+  showSessions = false;
 
-  @computed('session.currentRouteName')
-  get activeMenu() {
-    const { currentRouteName } = this.session;
-    if (currentRouteName === 'public.index' || currentRouteName === 'index') {
-      return 'Info';
-    } else if (currentRouteName === 'public.sessions') {
-      return 'Schedule';
-    } else if (currentRouteName === 'public.schedule') {
-      return 'Calendar';
-    } else if (currentRouteName === 'public.speakers') {
-      return 'Speakers';
-    } else if (currentRouteName === 'public.cfs.index') {
-      return 'Call for Speakers';
-    } else if (currentRouteName === 'public.coc') {
-      return 'Code of Conduct';
-    } else {
-      return 'Select section';
+  async didInsertElement() {
+    super.didInsertElement(...arguments);
+    const speakersCall = await this.event.speakersCall;
+    this.set('shouldShowCallforSpeakers',
+      speakersCall && speakersCall.announcement && (speakersCall.privacy === 'public'));
+
+    this.checkSpeakers();
+    this.checkSessions();
+  }
+
+  async checkSpeakers() {
+    this.showSpeakers = this.showSpeakers || (await this.loader.load(`/events/${this.event.id}/speakers?cache=true&public=true&fields[speaker]=id&page[size]=1&filter=${JSON.stringify(SPEAKERS_FILTER)}`)).data.length;
+  }
+
+  async checkSessions() {
+    const filters = [{
+      or: [
+        {
+          name : 'state',
+          op   : 'eq',
+          val  : 'confirmed'
+        },
+        {
+          name : 'state',
+          op   : 'eq',
+          val  : 'accepted'
+        }
+      ]
+    }];
+    this.showSessions = this.showSessions || (await this.loader.load(`/events/${this.event.id}/sessions?cache=true&public=true&fields[session]=id&page[size]=1&filter=${JSON.stringify(filters)}`)).data.length;
+  }
+
+  didRender() {
+    if (!this.activeSection) { return }
+    const target = document.querySelector(`[href='#${this.activeSection}']`);
+    if (target) {
+      // Delay click to give time to render
+      setTimeout(() => {
+        target.click();
+      }, 0);
     }
   }
-}
 
+  @action
+  goToSection(section) {
+    this.set('activeSection', section);
+    this.set('activeMenuSection', section);
+  }
+
+  @action
+  scrollToTarget(section) {
+    document.querySelector(`#${section}`).scrollIntoView({
+      behavior: 'smooth'
+    });
+    this.set('activeMenuSection', section);
+    this.set('activeSection', null);
+    document.querySelectorAll('.scroll').forEach(node => {
+      node.classList.remove('active');
+    });
+    document.querySelector(`[href='#${section}']`).classList.add('active');
+  }
+
+  @computed('event.schedulePublishedOn')
+  get isSchedulePublished() {
+    return this.event.schedulePublishedOn && this.event.schedulePublishedOn.toISOString() !== moment(0).toISOString();
+  }
+}
