+++ app/templates/public.hbs
@@ -25,17 +25,19 @@
       <div class="three wide column" id="public-event-content">
         {{#if this.device.isMobile}}
           {{#if this.displaySideMenu}}
-              <Public::SideMenu @event={{this.model}} />
+            <Public::SideMenuOuter @event={{this.model}} />
           {{/if}}
         {{else}}
           <div class="ui fluid side-menu">
-            <Public::SideMenu @event={{this.model}} />
+            <Public::SideMenuOuter @event={{this.model}} />
           </div>
         {{/if}}
       </div>
+      {{#unless this.device.isMobile}}
       <div class="ten wide column">
         {{outlet}}
       </div>
+      {{/unless}}
       <div class="three wide column">
         {{#if this.displaySideMenu}}
           <Public::AddToCalender @event={{this.model}} @location={{this.headerLocation}}/>
@@ -45,8 +47,16 @@
             </div>
           {{/if}}
           <Public::SocialLinks @externalUrl={{this.model.externalEventUrl}} @socialLinks={{this.model.socialLinks}}/>
+          {{#if (and this.device.isMobile this.twitterLink)}}
+              <div class="item">
+                <div class="twitter content">
+                  <Widgets::TwitterTimeline @handleOrProfile={{this.twitterLink.link}} />
+                </div>
+              </div>
+          {{/if}}
           {{#if (eq this.session.currentRouteName 'public.sessions')}}
-            <div class="ui basic segment">
+            <div class="{{if this.device.isMobile 'd-flex'}} ui basic segment">
+              <div class="mr-8">
               <h4>
                 {{t 'Tracks'}}
                 <LinkTo @route="public.sessions" @models={{array this.model.id}} @query={{hash track=null}} data-tooltip="{{t 'Clear Track Filter'}}">
@@ -62,6 +72,9 @@
                   <i class="circle icon" style={{css color=track.color}}></i> {{track.name}}
                 </LinkTo>
               {{/each}}
+              </div>
+              <br>
+              <div>
               <h4>
                 {{t 'Rooms'}}
                 <LinkTo @route="public.sessions" @models={{array this.model.id}} @query={{hash room=null}} data-tooltip="{{t 'Clear Room Filter'}}">
@@ -77,17 +90,23 @@
                   <i class="circle icon"></i> {{room.name}}
                 </LinkTo>
               {{/each}}
-            </div>
-          {{/if}}
-          {{#if this.twitterLink}}
-            <div class="item">
-              <div class="twitter content">
-                <Widgets::TwitterTimeline @handleOrProfile={{this.twitterLink.link}} />
               </div>
             </div>
           {{/if}}
+          {{#if (and (not this.device.isMobile) this.twitterLink)}}
+              <div class="item">
+                <div class="twitter content">
+                  <Widgets::TwitterTimeline @handleOrProfile={{this.twitterLink.link}} />
+                </div>
+              </div>
+          {{/if}}
         {{/if}}
       </div>
+      {{#if this.device.isMobile}}
+        <div class="ten wide column">
+          {{outlet}}
+        </div>
+      {{/if}}
     </div>
   </div>
 </div>
